+++ contracts/Power.sol
@@ -9,11 +9,11 @@ contract Power is ERC20Basic {
 
   string public name = "Acebusters Power";
   string public symbol = "ABP";
-  uint public decimals = 12;
+  uint256 public decimals = 12;
   
 
   // time it should take to power down
-  uint public downtime;
+  uint256 public downtime;
   // token contract address
   address nutzAddr;
   // sum of all outstanding power
@@ -21,7 +21,7 @@ contract Power is ERC20Basic {
   // authorized power
   uint256 authorizedPower = 0;
   // when powering down, at least totalSupply/minShare Power should be claimed
-  uint minShare = 10000;
+  uint256 minShare = 10000;
 
   // maxPower is a limit of total power that can be outstanding
   // maxPower has a valid value between outstandingPower and authorizedPow/2
@@ -33,13 +33,13 @@ contract Power is ERC20Basic {
   // data structure for withdrawals
   struct DownRequest {
     address owner;
-    uint total;
-    uint left;
-    uint start;
+    uint256 total;
+    uint256 left;
+    uint256 start;
   }
   DownRequest[] downs;
 
-  function Power(address _nutzAddr, uint _downtime) {
+  function Power(address _nutzAddr, uint256 _downtime) {
     nutzAddr = _nutzAddr;
     downtime = _downtime;
   }
@@ -58,7 +58,7 @@ contract Power is ERC20Basic {
     return authorizedPower;
   }
 
-  function vestedDown(uint _pos, uint _now) constant returns (uint256) {
+  function vestedDown(uint256 _pos, uint256 _now) constant returns (uint256) {
     if (downs.length <= _pos) {
       return 0;
     }
@@ -68,12 +68,12 @@ contract Power is ERC20Basic {
     // calculate amountVested
     // amountVested is amount that can be withdrawn according to time passed
     DownRequest req = downs[_pos];
-    uint timePassed = _now.sub(req.start);
+    uint256 timePassed = _now.sub(req.start);
     if (timePassed > downtime) {
      timePassed = downtime;
     }
-    uint amountVested = req.total.mul(timePassed).div(downtime);
-    uint amountFrozen = req.total.sub(amountVested);
+    uint256 amountVested = req.total.mul(timePassed).div(downtime);
+    uint256 amountFrozen = req.total.sub(amountVested);
     if (req.left <= amountFrozen) {
       return 0;
     }
@@ -89,23 +89,23 @@ contract Power is ERC20Basic {
   // ############################################
 
   // executes a powerdown request
-  function _downTick(uint _pos, uint _now) internal returns (bool success) {
-    uint amountPow = vestedDown(_pos, _now);
+  function _downTick(uint256 _pos, uint256 _now) internal returns (bool success) {
+    uint256 amountPow = vestedDown(_pos, _now);
     if (amountPow == 0) {
       throw;
     }
     DownRequest req = downs[_pos];
 
     // prevent power down in tiny steps
-    uint minStep = req.total.div(10);
+    uint256 minStep = req.total.div(10);
     if (amountPow < minStep && req.left > minStep) {
       throw;
     }
 
     // calculate token amount representing share of power
     var nutzContract = ERC20(nutzAddr);
-    uint totalBabz = nutzContract.totalSupply();
-    uint amountBabz = amountPow.mul(totalBabz).div(authorizedPower);
+    uint256 totalBabz = nutzContract.totalSupply();
+    uint256 amountBabz = amountPow.mul(totalBabz).div(authorizedPower);
     // transfer power and tokens
     outstandingPower = outstandingPower.sub(amountPow);
     req.left = req.left.sub(amountPow);
@@ -142,7 +142,7 @@ contract Power is ERC20Basic {
   }
 
   // this is called when NTZ are deposited into the burn pool
-  function dilutePower(uint _totalBabzBefore, uint _amountBabz) onlyNutzContract returns (bool) {
+  function dilutePower(uint256 _totalBabzBefore, uint256 _amountBabz) onlyNutzContract returns (bool) {
     if (authorizedPower == 0) {
       // during the first capital increase, set some big number as authorized shares
       authorizedPower = _totalBabzBefore.add(_amountBabz);
@@ -162,11 +162,11 @@ contract Power is ERC20Basic {
   // ############################################
 
   // this is called when NTZ are deposited into the power pool
-  function tokenFallback(address _from, uint _amountBabz, bytes32 _data) {
+  function tokenFallback(address _from, uint256 _amountBabz, bytes32 _data) {
     assert (msg.sender == nutzAddr);
     uint256 totalBabz = uint256(_data);
     assert(authorizedPower != 0 && _amountBabz != 0 && totalBabz != 0);
-    uint amountPow = _amountBabz.mul(authorizedPower).div(totalBabz);
+    uint256 amountPow = _amountBabz.mul(authorizedPower).div(totalBabz);
     // TODO: check amountPow is worth dealing with (not small percenage)
     // check pow limits
     assert(outstandingPower.add(amountPow) <= maxPower);
@@ -176,26 +176,26 @@ contract Power is ERC20Basic {
   }
 
   // registers a powerdown request
-  function transfer(address _to, uint _amountPower) returns (bool success) {
+  function transfer(address _to, uint256 _amountPower) returns (bool success) {
     // make Power not transferable
     assert(_to == nutzAddr);
     // prevent powering down tiny amounts
     assert(_amountPower >= authorizedPower.div(minShare));
 
     balances[msg.sender] = balances[msg.sender].sub(_amountPower);
-    uint pos = downs.length++;
+    uint256 pos = downs.length++;
     downs[pos] = DownRequest(msg.sender, _amountPower, _amountPower, now);
     return true;
   }
 
-  function downTick(uint _pos) returns (bool success) {
+  function downTick(uint256 _pos) returns (bool success) {
       return _downTick(_pos, now);
   }
 
   // !!!!!!!!!!!!!!!!!!!!!!!! IMPORTANT !!!!!!!!!!!!!!!!!!!!!
   // REMOVE THIS BEFORE DEPLOYMENT!!!!
   // needed for accelerated time testing
-  function downTickTest(uint _pos, uint _now) returns (bool success) {
+  function downTickTest(uint256 _pos, uint256 _now) returns (bool success) {
     return _downTick(_pos, _now);
   }
   // !!!!!!!!!!!!!!!!!!!!!!!! IMPORTANT !!!!!!!!!!!!!!!!!!!!!
